Show sample code for empty git hook script
When git hook is invoked with parameter it launches vim editor with given
parameter. The parameter means a kind of hook and usually it contains sample
code that has .sample as its suffix. If there's no hook script as given
parameter and there's a sample, show it on editor buffer for new hook script.