Apply syntax detecting to git hooks load sample hook.
In previous commit(fe77e9f696a83373cee46cb0d2bb91c7884d70dc) git hooks load
sample script when needed. But it just load the sample and doesn't apply syntax
highlighting in vim. There's a command that is "filetype detect" to apply syntax
highlight as per detected filetype. So, I'm applying it the to loaded sample
script that is not saved yet.